apps/mobile/backend/src/script/requirements.txt

fastapi
uvicorn
groq
python-dotenv
pydantic
